marketplace/requirements.txt
flask ~= 1.1
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
markupsafe ~=2.0.1
pytest ~= 5.4